restructure, refactor, and add an annotation
