+++ src/main/java/gov/va/schema/emis/vdrdodadapter/v2/DoDAdapterClient.java
@@ -4,10 +4,12 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.stereotype.Component;
 import org.springframework.ws.client.core.WebServiceTemplate;
 
 import javax.xml.bind.JAXBElement;
 
+@Component
 public class DoDAdapterClient {
   private static final Logger LOGGER = LoggerFactory.getLogger(DoDAdapterClient.class);
